requirements.txt
Flask==3.0.2
flask-mongoengine==1.0.0
python-dotenv==1.0.0
pandas==2.2.1
PyYAML==6.0.1
requests==2.31.0
openpyxl==3.1.2
flask-swagger-ui==4.11.1